+++ src/main/java/ru/yandex/practicum/filmorate/storage/inmemory/InMemoryFilmStorage.java
@@ -6,8 +6,8 @@ import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.utilities.KeyGenerator;
 
 @Repository
-public class InMemoryFilmStorage extends BaseInMemoryStorage<Film, Integer> {
-    public InMemoryFilmStorage(KeyGenerator<Integer> keyGenerator) {
+public class InMemoryFilmStorage extends BaseInMemoryStorage<Film, Long> {
+    public InMemoryFilmStorage(KeyGenerator<Long> keyGenerator) {
         super(keyGenerator);
     }
 }
