+++ src/main/java/ru/yandex/practicum/filmorate/storage/inmemory/InMemoryUserStorage.java
@@ -6,7 +6,7 @@ import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.utilities.IntegerIdGenerator;
 
 @Repository
-public class InMemoryUserStorage extends BaseInMemoryStorage<User, Integer> {
+public class InMemoryUserStorage extends BaseInMemoryStorage<User, Long> {
     public InMemoryUserStorage() {
         super(new IntegerIdGenerator());
     }
